+++ Cargo.toml
@@ -71,7 +71,6 @@ harness = false
 members = [
     "analyze",
     "ast",
-    "datagen",
     "error",
     "instrument",
     "lame",
