+++ ast/src/datagen/test_data.csv
@@ -1,0 +1,1 @@
+0.01,0.02,0.03
