--- datagen/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "datagen"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-csv = "1.1.6"
-num-rational = "0.3.2"
-serde = {version="1.0.68", features=["derive"]}
-serde_json = "1.0.68"
-weresocool_ast = { path = "../ast" }
-weresocool_shared = { path = "../shared" }
-scop = { path = "../scop" }
